pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-data-warehouse-skills"
version = "0.1.0"
description = "AI skills and utilities for StarRocks based data warehouse analysis"
requires-python = ">=3.10"
dependencies = ["pydantic>=2.6", "pyyaml>=6.0", "sqlparse>=0.5.0", "typer>=0.12.0", "rich>=13.7.0", "jinja2>=3.1.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
